requirements.txt
# Core Flask packages
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==2.3.7

# Form handling and validation
Flask-WTF==1.1.1
WTForms==3.0.1

# Database migrations
Flask-Migrate==4.0.5

# Email support
Flask-Mail==0.9.1

# Environment variable management
python-dotenv==1.0.0

# Production WSGI server
gunicorn==21.2.0

# Development tools
Flask-DebugToolbar==0.13.1

# Date and time handling
python-dateutil==2.8.2

# Password validation
bcrypt==4.0.1

# Pagination support
Flask-SQLAlchemy==3.0.5

# Testing
pytest==7.4.2
pytest-flask==1.2.0

# Security
Flask-Talisman==1.1.0

# API support (for future expansion)
Flask-RESTful==0.3.10
marshmallow==3.20.1
